BACKEND_ECOAGRIS/requirements.txt
asgiref
certifi
charset-normalizer
distlib
Django
django-cors-headers
django-filter
djangorestframework
djangorestframework-simplejwt
et-xmlfile
filelock
filter-and-pagination
idna
numpy
openpyxl
pandas
pillow
pipenv
platformdirs
PyJWT
PyMySQL
python-dateutil
pytz
requests
setuptools
six
sqlparse
tzdata
urllib3
virtualenv
wheel
